+++ my_class/checkers_board_class.py
